Fix for the memory leak when multiple batch runs are triggered (#807)
* Fix for the memory leak when multiple batch runs are triggered

The default glibc memory allocator is created memory fragmentations and is causing the off-heap memory of the application to grow indefinitly. This is replaced by the jemalloc allocator which avoids such problems.

Refer https://github.com/google/fhir-data-pipes/issues/777 for more details

* Fix for the memory leak when multiple batch runs are triggered

The default glibc memory allocator is created memory fragmentations and is causing the off-heap memory of the application to grow indefinitly. This is replaced by the jemalloc allocator which avoids such problems.

Refer https://github.com/google/fhir-data-pipes/issues/777 for more details

* comments refactored